refactor(init-config): removes superfluous function; touches up the typing a bit
