+++ test/cli/init-config/validators.spec.mjs
@@ -1,37 +1,5 @@
-/* eslint-disable no-unused-expressions */
 import { expect } from "chai";
-import validators from "../../../src/cli/init-config/inquirer-validators.js";
-
-describe("[U] cli/init-config/inquirer-validators - validateFileExistence", () => {
-  const WORKING_DIR = process.cwd();
-  const lFixturesDirectory =
-    "test/cli/init-config/__fixtures__/validate-file-existence";
-
-  beforeEach("set up", () => {
-    process.chdir(lFixturesDirectory);
-  });
-
-  afterEach("tear down", () => {
-    process.chdir(WORKING_DIR);
-  });
-
-  it("returns an error message when presented with an empty string", () => {
-    expect(validators.validateFileExistence("")).to.equal(
-      `hmm, '' doesn't seem to exist - could you try again?`
-    );
-  });
-
-  it("returns an error message when presented with a name of a non existing file", () => {
-    expect(validators.validateFileExistence("nope-does-not-exist")).to.equal(
-      `hmm, 'nope-does-not-exist' doesn't seem to exist - could you try again?`
-    );
-  });
-
-  it("returns true when presented with the name of a file that does exist", () => {
-    expect(validators.validateFileExistence("i-have-bytes-therefore-i-exist"))
-      .to.be.true;
-  });
-});
+import validators from "../../../src/cli/init-config/validators.js";
 
 describe("[U] cli/init-config/inquirer-validators - validateLocation", () => {
   const WORKING_DIR = process.cwd();
